try-elevenlabs/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
coloredlogs==15.0.1
elevenlabs==2.15.0
flatbuffers==25.2.10
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
joblib==1.5.2
mpmath==1.3.0
numpy==2.3.3
onnxruntime==1.22.1
openwakeword==0.6.0
packaging==25.0
protobuf==6.32.1
pvporcupine==3.0.5
PyAudio==0.2.14
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
